4th commit

+++ src/behavioral/design/patterns/strategy/pattern/CreditCardPayment.java
@@ -2,7 +2,7 @@ package behavioral.design.patterns.strategy.pattern;
 
 // Implement CreditCardPayment Strategy
 public class CreditCardPayment implements PaymentStrategy {
-    public void pay(int amount) {
+    public void pay(double amount) {
         System.out.println("Paid ₹" + amount + " using Credit Card.");
     }
 }
