+++ src/behavioral/design/patterns/strategy/pattern/PaymentStrategy.java
@@ -2,6 +2,6 @@ package behavioral.design.patterns.strategy.pattern;
 
 // Define the Strategy Interface
 public interface PaymentStrategy {
-    void pay(int amount);
+    void pay(double amount);
 }
